+++ src/plugins/nonebot_bison/config/db_config.py
@@ -1,7 +1,9 @@
 from collections import defaultdict
+from dataclasses import asdict
 from datetime import datetime, time
-from typing import Awaitable, Callable, Optional, Sequence
+from typing import Any, Awaitable, Callable, Optional, Sequence
 
+from nonebot import logger
 from nonebot_plugin_datastore import create_session
 from sqlalchemy import delete, func, select
 from sqlalchemy.exc import IntegrityError
