+++ src/plugins/nonebot_bison/subs_io/__init__.py
@@ -1,0 +1,3 @@
+from .subs_io import subscribes_export, subscribes_import
+
+__all__ = ["subscribes_export", "subscribes_import"]
