+++ src/plugins/nonebot_bison/subs_io/nbesf_model.py
@@ -1,0 +1,17 @@
+""" nbexf is Nonebot Bison Enchangable Subscribes File! """
+
+from dataclasses import dataclass
+
+from ..types import Category, Tag
+
+
+@dataclass
+class ItemModel:
+    user: int
+    user_type: str
+    target: str
+    target_name: str
+    platform_name: str
+    categories: list[Category]
+    tags: list[Tag]
+    # default_schedule_weight: int
