+++ src/plugins/nonebot_bison/subs_io/subs_io.py
@@ -1,88 +1,16 @@
 import json
 import time
-from dataclasses import asdict, dataclass
+from dataclasses import asdict
 from functools import partial
 from pathlib import Path
 
 from nonebot.log import logger
-from nonebot_bison.types import Category, Tag
-from nonebot_plugin_datastore.db import get_engine
 from objtyping import to_primitive
-from sqlalchemy.ext.asyncio.session import AsyncSession
 
-from .config_legacy import ConfigContent
-from .config_legacy import config as db_config_legacy
-from .config_legacy import drop
-from .db_config import SubscribeDupException
-from .db_config import config as db_config
-from .db_model import Subscribe, Target, User
-
-
-async def data_migrate():
-    if db_config_legacy.available:
-        logger.warning("You are still using legacy db, migrating to sqlite")
-        all_subs: list[ConfigContent] = list(
-            map(
-                lambda item: ConfigContent(**item),
-                db_config_legacy.get_all_subscribe().all(),
-            )
-        )
-        async with AsyncSession(get_engine()) as sess:
-            user_to_create = []
-            subscribe_to_create = []
-            platform_target_map: dict[str, tuple[Target, str, int]] = {}
-            for user in all_subs:
-                db_user = User(uid=user["user"], type=user["user_type"])
-                user_to_create.append(db_user)
-                user_sub_set = set()
-                for sub in user["subs"]:
-                    target = sub["target"]
-                    platform_name = sub["target_type"]
-                    target_name = sub["target_name"]
-                    key = f"{target}-{platform_name}"
-                    if key in user_sub_set:
-                        # a user subscribe a target twice
-                        logger.error(
-                            f"用户 {user['user_type']}-{user['user']} 订阅了 {platform_name}-{target_name} 两次，"
-                            "随机采用了一个订阅"
-                        )
-                        continue
-                    user_sub_set.add(key)
-                    if key in platform_target_map.keys():
-                        target_obj, ext_user_type, ext_user = platform_target_map[key]
-                        if target_obj.target_name != target_name:
-                            # GG
-                            logger.error(
-                                f"你的旧版本数据库中存在数据不一致问题，请完成迁移后执行重新添加{platform_name}平台的{target}"
-                                f"它的名字可能为{target_obj.target_name}或{target_name}"
-                            )
-
-                    else:
-                        target_obj = Target(
-                            platform_name=platform_name,
-                            target_name=target_name,
-                            target=target,
-                        )
-                        platform_target_map[key] = (
-                            target_obj,
-                            user["user_type"],
-                            user["user"],
-                        )
-                    subscribe_obj = Subscribe(
-                        user=db_user,
-                        target=target_obj,
-                        categories=sub["cats"],
-                        tags=sub["tags"],
-                    )
-                    subscribe_to_create.append(subscribe_obj)
-            sess.add_all(
-                user_to_create
-                + list(map(lambda x: x[0], platform_target_map.values()))
-                + subscribe_to_create
-            )
-            await sess.commit()
-            drop()
-            logger.info("migrate success")
+from ..config.db_config import SubscribeDupException
+from ..config.db_config import config as db_config
+from ..types import Category, Tag
+from .nbesf_model import ItemModel
 
 
 async def subscribes_export(export_path: Path = Path.cwd() / "data"):
@@ -155,18 +83,6 @@ async def subscribes_import(import_file_path: Path):
     参数:
         import_file_path (Path, optional): 需要导入的订阅文件的位置。
     """
-
-    @dataclass
-    class ItemModel:
-        user: int
-        user_type: str
-        target: str
-        target_name: str
-        platform_name: str
-        categories: list[Category]
-        tags: list[Tag]
-        # default_schedule_weight: int
-
     # 合法性检查与订阅包生成
     assert import_file_path.is_file()
 
@@ -181,8 +97,8 @@ async def subscribes_import(import_file_path: Path):
 
     assert isinstance(import_items, list)
 
+    sub_packs: list[ItemModel] = list()
     try:
-        sub_packs: list[ItemModel] = list()
         for item in import_items:
 
             assert isinstance(item, dict)
@@ -230,16 +146,16 @@ async def subscribes_import(import_file_path: Path):
     # 逐个添加订阅
     logger.info("检查结束，开始添加订阅流程")
 
-    if is_check_pass:
+    if is_check_pass and sub_packs:
 
         for sub_pack in sub_packs:
 
             try:
                 await db_config.add_subscribe(**asdict(sub_pack))
             except SubscribeDupException:
-                logger.warning(f"添加订阅条目 {repr(sub_pack)} 失败: 相同的订阅已存在")
+                logger.warning(f"！添加订阅条目 {repr(sub_pack)} 失败: 相同的订阅已存在")
             except Exception as e:
-                logger.error(f"添加订阅条目 {repr(sub_pack)} 失败: {repr(e)}")
+                logger.error(f"！添加订阅条目 {repr(sub_pack)} 失败: {repr(e)}")
             else:
                 logger.success(f"添加订阅条目 {repr(sub_pack)} 成功！")
